address changes from review
