+++ DQM/SiTrackerPhase2/plugins/BuildFile.xml
@@ -8,6 +8,7 @@
 <use name="Geometry/CommonDetUnit"/>
 <use name="Geometry/TrackerGeometryBuilder"/>
 <use name="DQMServices/Core"/>
+<use name="DQM/SiTrackerPhase2"/>
 <use name="boost_system"/>
 <use name="rootcling"/>
 <flags EDM_PLUGIN="1"/>
