+++ DQM/SiTrackerPhase2/plugins/Phase2ITMonitorRecHit.cc
@@ -10,31 +10,83 @@
 //
 // system include files
 #include <memory>
-#include "DQM/SiTrackerPhase2/plugins/Phase2ITMonitorRecHit.h"
-#include "DQM/SiTrackerPhase2/interface/TrackerPhase2DQMUtil.h"
-#include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include <map>
+#include <vector>
+#include <algorithm>
 #include "FWCore/Framework/interface/ESWatcher.h"
-#include <FWCore/ParameterSet/interface/ConfigurationDescriptions.h>
-#include <FWCore/ParameterSet/interface/ParameterSetDescription.h>
-#include <FWCore/ParameterSet/interface/ConfigurationDescriptions.h>
-#include <FWCore/ParameterSet/interface/ParameterSetDescription.h>
-#include "Geometry/Records/interface/TrackerDigiGeometryRecord.h"
-#include "Geometry/TrackerGeometryBuilder/interface/TrackerGeometry.h"
+#include "FWCore/Framework/interface/Frameworkfwd.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+#include "FWCore/Framework/interface/ESHandle.h"
+#include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
+#include "FWCore/ParameterSet/interface/ParameterSetDescription.h"
+#include "FWCore/ServiceRegistry/interface/Service.h"
+#include "FWCore/Utilities/interface/InputTag.h"
+#include "DataFormats/Common/interface/DetSetVector.h"
+#include "DataFormats/Common/interface/DetSetVectorNew.h"
+#include "DataFormats/Common/interface/Handle.h"
+#include "DataFormats/DetId/interface/DetId.h"
+#include "DataFormats/GeometrySurface/interface/LocalError.h"
+#include "DataFormats/GeometryVector/interface/LocalPoint.h"
+#include "DataFormats/TrackerCommon/interface/TrackerTopology.h"
+#include "DataFormats/TrackerRecHit2D/interface/SiPixelRecHitCollection.h"
 #include "Geometry/CommonDetUnit/interface/GeomDet.h"
 #include "Geometry/CommonDetUnit/interface/TrackerGeomDet.h"
 #include "Geometry/CommonDetUnit/interface/PixelGeomDetUnit.h"
 #include "Geometry/CommonDetUnit/interface/PixelGeomDetType.h"
 #include "Geometry/Records/interface/TrackerTopologyRcd.h"
 #include "Geometry/Records/interface/TrackerDigiGeometryRecord.h"
-#include "DataFormats/Common/interface/DetSetVector.h"
-#include "DataFormats/TrackerCommon/interface/TrackerTopology.h"
-#include "DataFormats/GeometrySurface/interface/LocalError.h"
-#include "DataFormats/GeometryVector/interface/LocalPoint.h"
+#include "Geometry/TrackerGeometryBuilder/interface/TrackerGeometry.h"
 // DQM Histograming
 #include "DQMServices/Core/interface/MonitorElement.h"
-//
-// constructors
-//
+#include "DQMServices/Core/interface/DQMEDAnalyzer.h"
+#include "DQMServices/Core/interface/DQMStore.h"
+#include "DQM/SiTrackerPhase2/interface/TrackerPhase2DQMUtil.h"
+
+class Phase2ITMonitorRecHit : public DQMEDAnalyzer {
+public:
+  explicit Phase2ITMonitorRecHit(const edm::ParameterSet&);
+  ~Phase2ITMonitorRecHit() override;
+  void bookHistograms(DQMStore::IBooker& ibooker, edm::Run const& iRun, edm::EventSetup const& iSetup) override;
+  void analyze(const edm::Event& iEvent, const edm::EventSetup& iSetup) override;
+  void dqmBeginRun(const edm::Run& iRun, const edm::EventSetup& iSetup) override;
+  static void fillDescriptions(edm::ConfigurationDescriptions& descriptions);
+
+private:
+  void fillITHistos(const edm::Event& iEvent);
+  void bookLayerHistos(DQMStore::IBooker& ibooker, unsigned int det_id, std::string& subdir);
+
+  edm::ParameterSet config_;
+  std::string geomType_;
+  const edm::EDGetTokenT<SiPixelRecHitCollection> tokenRecHitsIT_;
+  const edm::ESGetToken<TrackerGeometry, TrackerDigiGeometryRecord> geomToken_;
+  const edm::ESGetToken<TrackerTopology, TrackerTopologyRcd> topoToken_;
+  const TrackerGeometry* tkGeom_ = nullptr;
+  const TrackerTopology* tTopo_ = nullptr;
+  static constexpr float million = 1e6;
+  MonitorElement* numberRecHits_;
+  MonitorElement* globalXY_barrel_;
+  MonitorElement* globalXY_endcap_;
+  MonitorElement* globalRZ_barrel_;
+  MonitorElement* globalRZ_endcap_;
+
+  struct RecHitME {
+    MonitorElement* numberRecHits = nullptr;
+    MonitorElement* globalPosXY = nullptr;
+    MonitorElement* globalPosRZ = nullptr;
+    MonitorElement* localPosXY = nullptr;
+    MonitorElement* posX = nullptr;
+    MonitorElement* posY = nullptr;
+    MonitorElement* poserrX = nullptr;
+    MonitorElement* poserrY = nullptr;
+    MonitorElement* clusterSizeX = nullptr;
+    MonitorElement* clusterSizeY = nullptr;
+  };
+  std::map<std::string, RecHitME> layerMEs_;
+};
+
 Phase2ITMonitorRecHit::Phase2ITMonitorRecHit(const edm::ParameterSet& iConfig)
     : config_(iConfig),
       tokenRecHitsIT_(consumes<SiPixelRecHitCollection>(iConfig.getParameter<edm::InputTag>("rechitsSrc"))),
@@ -58,27 +110,25 @@ void Phase2ITMonitorRecHit::fillITHistos(const edm::Event& iEvent) {
   std::map<std::string, unsigned int> nrechitLayerMap;
   unsigned long int nTotrechitsinevt = 0;
   // Loop over modules
-  SiPixelRecHitCollection::const_iterator DSViter;
-  for (DSViter = rechits->begin(); DSViter != rechits->end(); ++DSViter) {
+  for (const auto& DSViter : *rechits) {
     // Get the detector id
-    unsigned int rawid(DSViter->detId());
+    unsigned int rawid(DSViter.detId());
     DetId detId(rawid);
     // Get the geomdet
     const GeomDetUnit* geomDetunit(tkGeom_->idToDetUnit(detId));
     if (!geomDetunit)
       continue;
-    std::string key = Phase2TkUtil::getITHistoId(detId.rawId(), tTopo_);
-    nTotrechitsinevt += DSViter->size();
+    std::string key = phase2tkutil::getITHistoId(detId.rawId(), tTopo_);
+    nTotrechitsinevt += DSViter.size();
     if (nrechitLayerMap.find(key) == nrechitLayerMap.end()) {
-      nrechitLayerMap.insert(std::make_pair(key, DSViter->size()));
+      nrechitLayerMap.emplace(key, DSViter.size());
     } else {
-      nrechitLayerMap[key] += DSViter->size();
+      nrechitLayerMap[key] += DSViter.size();
     }
 
-    edmNew::DetSet<SiPixelRecHit>::const_iterator rechitIt;
     //loop over rechits for a single detId
-    for (rechitIt = DSViter->begin(); rechitIt != DSViter->end(); ++rechitIt) {
-      LocalPoint lp = rechitIt->localPosition();
+    for (const auto& rechit : DSViter) {
+      LocalPoint lp = rechit.localPosition();
       Global3DPoint globalPos = geomDetunit->surface().toGlobal(lp);
       //in mm
       double gx = globalPos.x() * 10.;
@@ -95,9 +145,9 @@ void Phase2ITMonitorRecHit::fillITHistos(const edm::Event& iEvent) {
       }
       //layer wise histo
       if (layerMEs_[key].clusterSizeX)
-        layerMEs_[key].clusterSizeX->Fill(rechitIt->cluster()->sizeX());
+        layerMEs_[key].clusterSizeX->Fill(rechit.cluster()->sizeX());
       if (layerMEs_[key].clusterSizeY)
-        layerMEs_[key].clusterSizeY->Fill(rechitIt->cluster()->sizeY());
+        layerMEs_[key].clusterSizeY->Fill(rechit.cluster()->sizeY());
       if (layerMEs_[key].globalPosXY)
         layerMEs_[key].globalPosXY->Fill(gx, gy);
       if (layerMEs_[key].globalPosRZ)
@@ -110,16 +160,16 @@ void Phase2ITMonitorRecHit::fillITHistos(const edm::Event& iEvent) {
         layerMEs_[key].posY->Fill(lp.y());
       float eta = geomDetunit->surface().toGlobal(lp).eta();
       if (layerMEs_[key].poserrX)
-        layerMEs_[key].poserrX->Fill(eta, million * rechitIt->localPositionError().xx());
+        layerMEs_[key].poserrX->Fill(eta, million * rechit.localPositionError().xx());
       if (layerMEs_[key].poserrY)
-        layerMEs_[key].poserrY->Fill(eta, million * rechitIt->localPositionError().yy());
+        layerMEs_[key].poserrY->Fill(eta, million * rechit.localPositionError().yy());
     }  //end loop over rechits of a detId
   }    //End loop over DetSetVector
 
   //fill nRecHits per event
   numberRecHits_->Fill(nTotrechitsinevt);
   //fill nRecHit counter per layer
-  for (auto& lme : nrechitLayerMap)
+  for (const auto& lme : nrechitLayerMap)
     if (layerMEs_[lme.first].numberRecHits)
       layerMEs_[lme.first].numberRecHits->Fill(lme.second);
 }
@@ -143,23 +193,23 @@ void Phase2ITMonitorRecHit::bookHistograms(DQMStore::IBooker& ibooker,
   //Global histos for IT
   HistoName.str("");
   HistoName << "NumberRecHits";
-  numberRecHits_ = Phase2TkUtil::book1DFromPSet(
+  numberRecHits_ = phase2tkutil::book1DFromPSet(
       config_.getParameter<edm::ParameterSet>("GlobalNumberRecHits"), HistoName.str(), ibooker);
   HistoName.str("");
   HistoName << "Global_Position_XY_IT_barrel";
-  globalXY_barrel_ = Phase2TkUtil::book2DFromPSet(
+  globalXY_barrel_ = phase2tkutil::book2DFromPSet(
       config_.getParameter<edm::ParameterSet>("GlobalPositionXY_PXB"), HistoName.str(), ibooker);
   HistoName.str("");
   HistoName << "Global_Position_RZ_IT_barrel";
-  globalRZ_barrel_ = Phase2TkUtil::book2DFromPSet(
+  globalRZ_barrel_ = phase2tkutil::book2DFromPSet(
       config_.getParameter<edm::ParameterSet>("GlobalPositionRZ_PXB"), HistoName.str(), ibooker);
   HistoName.str("");
   HistoName << "Global_Position_XY_IT_endcap";
-  globalXY_endcap_ = Phase2TkUtil::book2DFromPSet(
+  globalXY_endcap_ = phase2tkutil::book2DFromPSet(
       config_.getParameter<edm::ParameterSet>("GlobalPositionXY_PXEC"), HistoName.str(), ibooker);
   HistoName.str("");
   HistoName << "Global_Position_RZ_IT_endcap";
-  globalRZ_endcap_ = Phase2TkUtil::book2DFromPSet(
+  globalRZ_endcap_ = phase2tkutil::book2DFromPSet(
       config_.getParameter<edm::ParameterSet>("GlobalPositionRZ_PXEC"), HistoName.str(), ibooker);
 
   //Now book layer wise histos
@@ -173,14 +223,14 @@ void Phase2ITMonitorRecHit::bookHistograms(DQMStore::IBooker& ibooker,
         continue;
       unsigned int detId_raw = det_u->geographicalId().rawId();
       edm::LogInfo("Phase2ITMonitorRecHit") << "Detid:" << detId_raw << "\tsubdet=" << det_u->subDetector()
-                                            << "\t key=" << Phase2TkUtil::getITHistoId(detId_raw, tTopo_) << std::endl;
+                                            << "\t key=" << phase2tkutil::getITHistoId(detId_raw, tTopo_) << std::endl;
       bookLayerHistos(ibooker, detId_raw, dir);
     }
   }
 }
 // -- Book Layer Histograms
 void Phase2ITMonitorRecHit::bookLayerHistos(DQMStore::IBooker& ibooker, unsigned int det_id, std::string& subdir) {
-  std::string key = Phase2TkUtil::getITHistoId(det_id, tTopo_);
+  std::string key = phase2tkutil::getITHistoId(det_id, tTopo_);
   if (key.empty())
     return;
   if (layerMEs_.find(key) == layerMEs_.end()) {
@@ -191,54 +241,54 @@ void Phase2ITMonitorRecHit::bookLayerHistos(DQMStore::IBooker& ibooker, unsigned
     edm::LogInfo("Phase2ITMonitorRecHit") << " Booking Histograms in : " << (subdir + "/" + key);
     histoName.str("");
     histoName << "Number_RecHits";
-    local_histos.numberRecHits = Phase2TkUtil::book1DFromPSet(
+    local_histos.numberRecHits = phase2tkutil::book1DFromPSet(
         config_.getParameter<edm::ParameterSet>("LocalNumberRecHits"), histoName.str(), ibooker);
 
     histoName.str("");
     histoName << "RecHit_X";
     local_histos.posX =
-        Phase2TkUtil::book1DFromPSet(config_.getParameter<edm::ParameterSet>("RecHitPosX"), histoName.str(), ibooker);
+        phase2tkutil::book1DFromPSet(config_.getParameter<edm::ParameterSet>("RecHitPosX"), histoName.str(), ibooker);
 
     histoName.str("");
     histoName << "RecHit_Y";
     local_histos.posY =
-        Phase2TkUtil::book1DFromPSet(config_.getParameter<edm::ParameterSet>("RecHitPosY"), histoName.str(), ibooker);
+        phase2tkutil::book1DFromPSet(config_.getParameter<edm::ParameterSet>("RecHitPosY"), histoName.str(), ibooker);
 
     histoName.str("");
     histoName << "RecHit_X_error_Vs_eta";
-    local_histos.poserrX = Phase2TkUtil::bookProfile1DFromPSet(
+    local_histos.poserrX = phase2tkutil::bookProfile1DFromPSet(
         config_.getParameter<edm::ParameterSet>("RecHitPosErrorX_Eta"), histoName.str(), ibooker);
 
     histoName.str("");
     histoName << "RecHit_Y_error_Vs_eta";
-    local_histos.poserrY = Phase2TkUtil::bookProfile1DFromPSet(
+    local_histos.poserrY = phase2tkutil::bookProfile1DFromPSet(
         config_.getParameter<edm::ParameterSet>("RecHitPosErrorY_Eta"), histoName.str(), ibooker);
 
     histoName.str("");
     histoName << "Cluster_SizeX";
-    local_histos.clusterSizeX = Phase2TkUtil::book1DFromPSet(
+    local_histos.clusterSizeX = phase2tkutil::book1DFromPSet(
         config_.getParameter<edm::ParameterSet>("LocalClusterSizeX"), histoName.str(), ibooker);
 
     histoName.str("");
     histoName << "Cluster_SizeY";
-    local_histos.clusterSizeY = Phase2TkUtil::book1DFromPSet(
+    local_histos.clusterSizeY = phase2tkutil::book1DFromPSet(
         config_.getParameter<edm::ParameterSet>("LocalClusterSizeY"), histoName.str(), ibooker);
 
     histoName.str("");
     histoName << "Global_Position_XY";
-    local_histos.globalPosXY = Phase2TkUtil::book2DFromPSet(
+    local_histos.globalPosXY = phase2tkutil::book2DFromPSet(
         config_.getParameter<edm::ParameterSet>("GlobalPositionXY_perlayer"), histoName.str(), ibooker);
 
     histoName.str("");
     histoName << "Global_Position_RZ";
-    local_histos.globalPosRZ = Phase2TkUtil::book2DFromPSet(
+    local_histos.globalPosRZ = phase2tkutil::book2DFromPSet(
         config_.getParameter<edm::ParameterSet>("GlobalPositionRZ_perlayer"), histoName.str(), ibooker);
 
     histoName.str("");
     histoName << "Local_Position_XY";
-    local_histos.localPosXY = Phase2TkUtil::book2DFromPSet(
+    local_histos.localPosXY = phase2tkutil::book2DFromPSet(
         config_.getParameter<edm::ParameterSet>("LocalPositionXY"), histoName.str(), ibooker);
-    layerMEs_.insert(std::make_pair(key, local_histos));
+    layerMEs_.emplace(key, local_histos);
   }
 }
