+++ Validation/SiTrackerPhase2V/interface/TrackerPhase2ValidationUtil.h
@@ -2,22 +2,10 @@
 #define _Validation_SiTrackerPhase2V_TrackerPhase2ValidationUtil_h
 #include "SimDataFormats/TrackingHit/interface/PSimHit.h"
 #include "SimDataFormats/Track/interface/SimTrack.h"
-#include <string>
-#include <sstream>
 
-namespace Phase2TkUtil {
+namespace phase2tkutil {
 
-  bool isPrimary(const SimTrack& simTrk, const PSimHit* simHit) {
-    bool retval = false;
-    unsigned int trkId = simTrk.trackId();
-    if (trkId != simHit->trackId())
-      return retval;
-    int vtxIndex = simTrk.vertIndex();
-    int ptype = simHit->processType();
-    if ((vtxIndex == 0) && (ptype == 0))
-      return true;
-    return false;
-  }
+  bool isPrimary(const SimTrack& simTrk, const PSimHit* simHit);
 
-}  // namespace Phase2TkUtil
+}  // namespace phase2tkutil
 #endif
