+++ Validation/SiTrackerPhase2V/plugins/BuildFile.xml
@@ -20,7 +20,8 @@
 <use name="Geometry/TrackerGeometryBuilder"/>
 <use name="Geometry/CommonDetUnit"/>
 <use name="DQMServices/Core"/>
+<use name="DQM/SiTrackerPhase2"/>
+<use name="Validation/SiTrackerPhase2V"/>
 <use name="boost_system"/>
 <use name="rootcling"/>
-<flags EDM_PLUGIN="1"/>
 <flags EDM_PLUGIN="1"/>
