+++ Validation/SiTrackerPhase2V/src/BuildFile.xml
@@ -1,0 +1,5 @@
+<use name="SimDataFormats/Track"/>
+<use name="SimDataFormats/TrackingHit"/>
+<export>
+  <lib name="1"/>
+</export>
